requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
Pillow==8.3.1
werkzeug==3.0.3
zipp==3.18.1
